infra/requirements.txt
# AWS CDK Core
aws-cdk-lib==2.100.0
constructs>=10.0.0,<11.0.0

# AWS CDK Alpha modules (for experimental features)
aws-cdk.aws-lambda-python-alpha==2.100.0a0
aws-cdk.aws-apigatewayv2-alpha==2.100.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.100.0a0

# Python dependencies
boto3>=1.26.0
botocore>=1.29.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
moto>=4.0.0

# Development dependencies
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
isort>=5.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0